# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditActorDto": {
        "properties": {
          "id": {
            "example": "user_2abc",
            "type": "string"
          },
          "type": {
            "enum": [
              "operator",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "type"
        ],
        "type": "object"
      },
      "AuditEventDto": {
        "properties": {
          "action": {
            "example": "organization.membership.updated",
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/AuditActorDto"
          },
          "context": {
            "additionalProperties": true,
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          },
          "target": {
            "$ref": "#/components/schemas/AuditTargetDto"
          }
        },
        "required": [
          "action",
          "actor",
          "context",
          "id",
          "occurredAt",
          "target"
        ],
        "type": "object"
      },
      "AuditEventListDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventDto"
            },
            "type": "array"
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfoDto"
          }
        },
        "required": [
          "items",
          "pageInfo"
        ],
        "type": "object"
      },
      "AuditTargetDto": {
        "properties": {
          "id": {
            "example": "user_3def",
            "type": "string"
          },
          "type": {
            "example": "membership",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type"
        ],
        "type": "object"
      },
...
      "PageInfoDto": {
        "properties": {
          "hasNextPage": {
            "example": true,
            "type": "boolean"
          },
          "nextCursor": {
            "description": "Opaque cursor for the next page, or null at the end.",
            "example": "eyJ2IjoxLCJzY29wZSI6ImNvbGxlY3Rpb24ifQ",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "hasNextPage",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "correlationId": {
            "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6",
            "type": "string"
          },
          "detail": {
            "example": "One or more request values are invalid.",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemValidationErrorDto"
            },
            "type": "array"
          },
          "instance": {
            "example": "urn:uuid:f81d4fae-7dec-11d0-a765-00a0c91e6bf6",
            "type": "string"
          },
          "status": {
            "example": 400,
            "type": "number"
          },
          "title": {
            "example": "Request validation failed",
            "type": "string"
          },
          "type": {
            "example": "urn:problem:next-nest-saas-starter:validation",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "correlationId"
        ],
        "type": "object"
      },
      "ProblemValidationErrorDto": {
        "properties": {
          "detail": {
            "example": "limit must not be greater than 100",
            "type": "string"
          },
          "pointer": {
            "example": "#/query/limit",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "pointer"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "clerk-session": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Stable HTTP contract for the SaaS foundation.",
    "title": "Next Nest SaaS Starter API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}/audit-events": {
      "get": {
        "operationId": "listOrganizationAuditEvents",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor returned by the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventListDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "clerk-session": []
          }
        ],
        "summary": "",
        "tags": [
          "audit-events"
        ]
      }
    },
...
  },
  "servers": [],
...
}